In this part, we implement the flow of verifier user's JWT token, then make an authorized user to access subreddit resource.
As the Spring security uses filter to do such authentication, we first we have a JwtAuthenticationFilter, this class extends OncePerRequestFilter to fulfil its filter property.

The verification action takes place at JwtProvider.

JWT is signed by private key then send to user, and verififed by publicKey

the validateToken method on JwtProvider will verify token, and if pass the JwtAuthenticationFilter will append the authenticationToken to SecurityContextHolder, it makes this user able to access the data. Last don't forget the filterChain to pass this connection.

SubredditController, SubredditDto, and SubredditService are created to complete users' requests to subreddit.

SubredditController as a Rest Controller to handle the GET, POST request, and invoke the corresponding service at SubredditService.

SubredditDto wrap the the data from DB to obj for user to see.

SubredditService uses SubredditRepository obj to access DB, there are two mapping functions transfer SubReddit Obj to SubredditDto and vice versa.