As the last commit we discussed the two mapping functions in SubredditService, they can be replaced by MapStruct tool.
Here we have a mapper package for this MapStruct. We only have to define a interface and it maps the fields on DB Obj and Dto obj automatically, we only have to wire some fields manually if they are not pair. In this case is the numberOfPosts in subredditDto obj, we have annotation @Mapping(target = "numberOfPosts", expression = "java(mapPosts(subreddit.getPosts()))") for this. The expression related to the function to retrieve the target value, which can be customized.

MapStruct tool required Maven plugin support, which configured in pom.xml file, and the implementation will be created while building.