In this Commit, the whole PostController PostService and related PostDao are fully connected and working.
The flow is similar with the subreddit flow, basically we used the RestController to map the url, handle the request by corresponding services.

The data store and get still use the MapStruct which is PostMapper in this case. It generated the DB to Dto and Dto to DB methods easy for data transferring.