In this Commit, the whole CommentsController, CommentService and CommentMapper are fully connected and working.
The flow is similar with the subreddit, and post flow, basically we used the RestController to map the url, handle the request by corresponding services.

The data store and get still use the MapStruct which is CommentMapper in this case. It generated the DB to Dto and Dto to DB methods easy for data transferring.