In this comment, we implements complete logout and maintain user login status mechanisms
First we have to understand that the Jwt token is a stateless authentication mechanism, which means we don't store this token inside DB, but use public key to verify it. In this case we have a small condition for token which is the expiration time for each token, it can make sure this token expire after certain time, and it's the only way for now to invalid a token. Since expiring is the only way, we want to make sure the expiration time is as quick as possible, we made it in 15 minutes. However here is a problem, if a token expires in 15 minutes, users will not happy for re-login 15 minutes, it's time to implement a "refresh token" pair with expire time for this to work.

RefreshToken is a UUID string token user can be received with Jwt token when they login, and this token should be store to DB for validation since user can use this RefreshToken to generate a new Jwt token, and after user logout, this RefreshToken is removed from DB, so user cannot refresh the jwt token anymore, and let the jwt token expire to be invalided.

In fact this implementation can leave a security vulnerability, which the time we chose to logout is not really the time this jwt token invalided! user can still use this token to access resources til it expires.

Here is the new flow. AuthenticationResponse is changed to have two more fields called "expiresAt" and "refreshToken" so user can receive them while login. We added all RefreshToken model, repo and responseDto for data transferring. So when AuthController received the RefreshToken request, it will check if RefreshToken is inside DB, and then uses the username to generate a new jwt token for user, and send back. The logout method will simply delete the refresh token from DB to provent refresh Jwt again.